Remove Dead Key preview
With a keyboard layout such as "US-INTL", dead keys are a way to display
accented characters (such as "á"), by typing an accent, followed by a
letter. By default, my desktop environment shows a "preview" of the
accent (for example, when I press ' , it will show a underlined ´ ,
before I even type my letter). This, for some reason causes problems
with the vscode neovim extension, more specifically with macros (I can't
type a standalone ' when creating a macro). This commit this dead key
preview, to fix that.