Load .xprofile in wayland as well
The file .xprofile will be used to store any configuration related
specifically to graphic sessions (such as wayland and X11), which is why
I moved the changes made in the commit 869b08b there). The problem is,
this file is not opened by default in wayland sessions, only X11 ones.
With this commit, these files will be opened in wayland sessions as
well.